% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{create_pipeline}
\alias{create_pipeline}
\title{Pipeline}
\usage{
create_pipeline(...)
}
\description{
Create a pipeline.
}
\details{
\itemize{
  \item{\code{lang}: Model language.}
  \item{\code{treebank}: Treebank code from \url{https://stanfordnlp.github.io/stanfordnlp/models.html}..}
}
}
\examples{
\dontrun{
nlp <- create_pipeline()
}

}
