% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{download_model}
\alias{download_model}
\title{Download model}
\usage{
download_model(lang = "en", dir = NULL)
}
\arguments{
\item{lang}{Model language.}

\item{dir}{Path to directory where to save model.}
}
\description{
Download a language pack.
}
\examples{
\dontrun{
download_model()
}

}
