% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{create_pipeline}
\alias{create_pipeline}
\title{Pipeline}
\usage{
create_pipeline(...)
}
\arguments{
\item{...}{Parameters (all optional), see details.}
}
\description{
Create a pipeline.
}
\details{
\itemize{
  \item{\code{lang}: Model language.}
  \item{\code{treebank}: Treebank code from \url{https://stanfordnlp.github.io/stanfordnlp/models.html}.}
  \item{\code{processors}: Processors to use \code{tokenize,mwt,pos,lemma,depparse}.}
  \item{\code{models_dir}: Directory where model is stored.}
  \item{\code{use_gpu}: Boolean indicating whether to use the machine's GPU.}
}
}
\examples{
\dontrun{
nlp <- create_pipeline()
doc <- nlp("Barack Obama was born in Hawaii.")
doc$sentences[[0]]$print_dependencies()
}

}
