% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{download_model}
\alias{download_model}
\title{Download model}
\usage{
download_model(lang = "en")
}
\arguments{
\item{lang}{Model language.}
}
\description{
Download a language pack.
}
\examples{
\dontrun{
download_model()
}

}
